+++ src/os_cloud_nodelet.cpp
@@ -20,21 +20,18 @@
 #include <sensor_msgs/Imu.h>
 #include <sensor_msgs/PointCloud2.h>
 
-#include <algorithm>
-#include <chrono>
-#include <memory>
-#include <cassert>
-
 #include "ouster_ros/PacketMsg.h"
 #include "os_transforms_broadcaster.h"
 #include "imu_packet_handler.h"
 #include "lidar_packet_handler.h"
 #include "point_cloud_processor.h"
 #include "laser_scan_processor.h"
-
-namespace sensor = ouster::sensor;
+#include "point_cloud_processor_factory.h"
 
 namespace ouster_ros {
+
+namespace sensor = ouster::sensor;
+using ouster_ros::PacketMsg;
 
 class OusterCloud : public nodelet::Nodelet {
    public:
@@ -91,14 +88,14 @@ class OusterCloud : public nodelet::Nodelet {
     void create_publishers_subscribers(const sensor::sensor_info& info) {
         auto& pnh = getPrivateNodeHandle();
         auto proc_mask = pnh.param("proc_mask", std::string{"IMU|PCL|SCAN"});
-        auto tokens = parse_tokens(proc_mask, '|');
+        auto tokens = impl::parse_tokens(proc_mask, '|');
 
         auto timestamp_mode = pnh.param("timestamp_mode", std::string{});
         double ptp_utc_tai_offset = pnh.param("ptp_utc_tai_offset", -37.0);
 
         auto& nh = getNodeHandle();
 
-        if (check_token(tokens, "IMU")) {
+        if (impl::check_token(tokens, "IMU")) {
             imu_pub = nh.advertise<sensor_msgs::Imu>("imu", 100);
             imu_packet_handler = ImuPacketHandler::create_handler(
                 info, tf_bcast.imu_frame_id(), timestamp_mode,
@@ -115,41 +112,52 @@ class OusterCloud : public nodelet::Nodelet {
         int num_returns = get_n_returns(info);
 
         std::vector<LidarScanProcessor> processors;
-        if (check_token(tokens, "PCL")) {
+        if (impl::check_token(tokens, "PCL")) {
             lidar_pubs.resize(num_returns);
             for (int i = 0; i < num_returns; ++i) {
                 lidar_pubs[i] = nh.advertise<sensor_msgs::PointCloud2>(
                     topic_for_return("points", i), 10);
             }
 
-            processors.push_back(PointCloudProcessor::create(
-                info, tf_bcast.point_cloud_frame_id(),
-                tf_bcast.apply_lidar_to_sensor_transform(),
-                [this](PointCloudProcessor::OutputType msgs) {
-                    for (size_t i = 0; i < msgs.size(); ++i) {
-                        if (msgs[i]->header.stamp > last_msg_ts)
-                            last_msg_ts = msgs[i]->header.stamp;
-                        lidar_pubs[i].publish(*msgs[i]);
+            auto point_type = pnh.param("point_type", std::string{"original"});
+            processors.push_back(
+                PointCloudProcessorFactory::create_point_cloud_processor(point_type,
+                    info, tf_bcast.point_cloud_frame_id(),
+                    tf_bcast.apply_lidar_to_sensor_transform(),
+                    [this](PointCloudProcessor_OutputType msgs) {
+                        for (size_t i = 0; i < msgs.size(); ++i) {
+                            if (msgs[i]->header.stamp > last_msg_ts)
+                                last_msg_ts = msgs[i]->header.stamp;
+                            lidar_pubs[i].publish(*msgs[i]);
+                        }
                     }
-                }));
+                )
+            );
+
+            // warn about profile incompatibility
+            if (PointCloudProcessorFactory::point_type_requires_intensity(point_type) &&
+                info.format.udp_profile_lidar == UDPProfileLidar::PROFILE_RNG15_RFL8_NIR8) {
+                NODELET_WARN_STREAM(
+                    "selected point type '" << point_type << "' is not compatible with the current udp profile: RNG15_RFL8_NIR8");
+            }
         }
 
-        if (check_token(tokens, "SCAN")) {
+        if (impl::check_token(tokens, "SCAN")) {
             scan_pubs.resize(num_returns);
             for (int i = 0; i < num_returns; ++i) {
                 scan_pubs[i] = nh.advertise<sensor_msgs::LaserScan>(
                     topic_for_return("scan", i), 10);
             }
 
-            // TODO: avoid duplication in os_cloud_node
+            // TODO: avoid this duplication in os_cloud_node
             int beams_count = static_cast<int>(get_beams_count(info));
             int scan_ring = pnh.param("scan_ring", 0);
             scan_ring = std::min(std::max(scan_ring, 0), beams_count - 1);
             if (scan_ring != pnh.param("scan_ring", 0)) {
                 NODELET_WARN_STREAM(
-                    "scan ring is set to a value that exceeds available range"
-                    "please choose a value between [0, " << beams_count <<
-                    "], ring value clamped to: " << scan_ring);
+                    "scan ring is set to a value that exceeds available range "
+                    "please choose a value between [0, " << beams_count << "], "
+                    "ring value clamped to: " << scan_ring);
             }
 
             processors.push_back(LaserScanProcessor::create(
@@ -163,7 +171,7 @@ class OusterCloud : public nodelet::Nodelet {
                 }));
         }
 
-        if (check_token(tokens, "PCL") || check_token(tokens, "SCAN")) {
+        if (impl::check_token(tokens, "PCL") || impl::check_token(tokens, "SCAN")) {
             lidar_packet_handler = LidarPacketHandler::create_handler(
                 info, processors, timestamp_mode,
                 static_cast<int64_t>(ptp_utc_tai_offset * 1e+9));