+++ src/point_cloud_processor.h
@@ -14,25 +14,38 @@
 #include "ouster_ros/os_ros.h"
 // clang-format on
 
+#include "point_cloud_compose.h"
 #include "lidar_packet_handler.h"
 
 namespace ouster_ros {
 
+// Moved out of PointCloudProcessor to avoid type templatization
+using PointCloudProcessor_OutputType =
+    std::vector<std::shared_ptr<sensor_msgs::PointCloud2>>;
+using PointCloudProcessor_PostProcessingFn = std::function<void(PointCloudProcessor_OutputType)>;
+
+
+template <class PointT>
 class PointCloudProcessor {
    public:
-    using OutputType = std::vector<std::shared_ptr<sensor_msgs::PointCloud2>>;
-    using PostProcessingFn = std::function<void(OutputType)>;
+    using ScanToCloudFn = std::function<void(ouster_ros::Cloud<PointT>& cloud,
+                                        const ouster::PointsF& points,
+                                        uint64_t scan_ts, const ouster::LidarScan& ls,
+                                        const std::vector<int>& pixel_shift_by_row,
+                                        int return_index)>;
 
    public:
     PointCloudProcessor(const ouster::sensor::sensor_info& info,
                         const std::string& frame_id,
                         bool apply_lidar_to_sensor_transform,
-                        PostProcessingFn func)
+                        ScanToCloudFn scan_to_cloud_fn_,
+                        PointCloudProcessor_PostProcessingFn post_processing_fn_)
         : frame(frame_id),
           pixel_shift_by_row(info.format.pixel_shift_by_row),
           cloud{info.format.columns_per_frame, info.format.pixels_per_column},
           pc_msgs(get_n_returns(info)),
-          post_processing_fn(func) {
+          scan_to_cloud_fn(scan_to_cloud_fn_),
+          post_processing_fn(post_processing_fn_) {
         for (size_t i = 0; i < pc_msgs.size(); ++i)
             pc_msgs[i] = std::make_shared<sensor_msgs::PointCloud2>();
         ouster::mat4d additional_transform =
@@ -52,7 +65,8 @@ class PointCloudProcessor {
     }
 
    private:
-    void pcl_toROSMsg(const ouster_ros::Cloud& pcl_cloud,
+    template <typename T>
+    void pcl_toROSMsg(const ouster_ros::Cloud<T>& pcl_cloud,
                       sensor_msgs::PointCloud2& cloud) {
         // TODO: remove the staging step in the future
         pcl::toPCLPointCloud2(pcl_cloud, staging_pcl_pc2);
@@ -62,9 +76,13 @@ class PointCloudProcessor {
     void process(const ouster::LidarScan& lidar_scan, uint64_t scan_ts,
                  const ros::Time& msg_ts) {
         for (int i = 0; i < static_cast<int>(pc_msgs.size()); ++i) {
-            scan_to_cloud_f_destaggered(cloud, points, lut_direction,
-                                        lut_offset, scan_ts, lidar_scan,
+            auto range_channel = static_cast<sensor::ChanField>(sensor::ChanField::RANGE + i);
+            auto range = lidar_scan.field<uint32_t>(range_channel);
+            ouster::cartesianT(points, range, lut_direction, lut_offset);
+
+            scan_to_cloud_fn(cloud, points, scan_ts, lidar_scan,
                                         pixel_shift_by_row, i);
+
             pcl_toROSMsg(cloud, *pc_msgs[i]);
             pc_msgs[i]->header.stamp = msg_ts;
             pc_msgs[i]->header.frame_id = frame;
@@ -77,9 +95,10 @@ class PointCloudProcessor {
     static LidarScanProcessor create(const ouster::sensor::sensor_info& info,
                                      const std::string& frame,
                                      bool apply_lidar_to_sensor_transform,
-                                     PostProcessingFn func) {
+                                     ScanToCloudFn scan_to_cloud_fn_,
+                                     PointCloudProcessor_PostProcessingFn post_processing_fn) {
         auto handler = std::make_shared<PointCloudProcessor>(
-            info, frame, apply_lidar_to_sensor_transform, func);
+            info, frame, apply_lidar_to_sensor_transform, scan_to_cloud_fn_, post_processing_fn);
 
         return [handler](const ouster::LidarScan& lidar_scan, uint64_t scan_ts,
                          const ros::Time& msg_ts) {
@@ -98,11 +117,10 @@ class PointCloudProcessor {
     ouster::PointsF lut_offset;
     ouster::PointsF points;
     std::vector<int> pixel_shift_by_row;
-    ouster_ros::Cloud cloud;
-
-    OutputType pc_msgs;
-
-    PostProcessingFn post_processing_fn;
+    ouster_ros::Cloud<PointT> cloud;
+    PointCloudProcessor_OutputType pc_msgs;
+    ScanToCloudFn scan_to_cloud_fn;
+    PointCloudProcessor_PostProcessingFn post_processing_fn;
 };
 
 }  // namespace ouster_ros